requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
biopython==1.87
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.4
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
executing==2.2.1
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gemmi==0.7.5
graphql-core==3.2.8
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
rcsb-api==1.7.1
requests==2.33.1
rich==15.0.0
rustworkx==0.17.1
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.23.0
